requirements.txt
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
requests>=2.28.0

# Optional: For Excel export functionality
pandas>=1.5.0
openpyxl>=3.0.0